[UPDATE] code review guardar db
